Django API/GradeEZ API/requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
catboost==0.23.2
click==7.1.2
cycler==0.10.0
Django==3.0.8
djangorestframework==3.11.0
google==3.0.0
graphviz==0.14.1
joblib==0.16.0
kiwisolver==1.2.0
language-check==1.1
lightgbm==2.3.1
matplotlib==3.3.0
nltk==3.5
numpy==1.19.1
pandas==1.0.5
Pillow==7.2.0
plotly==4.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
retrying==1.3.3
scikit-learn==0.23.1
scipy==1.5.2
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
threadpoolctl==2.1.0
tqdm==4.48.0
urllib3==1.25.10
